[Qt] Fix build of WTR and the QML2 plugin on Windows https://bugs.webkit.org/show_bug.cgi?id=96652
Reviewed by Tor Arne Vestbø.

Some components that we build such as WTR or the QML2 plugin include
WebKit2 header files as well as WebKit2's config.h. Only the include
paths of WebKit2 are required for that at the moment (WEBKIT +=
webkit2). Only on Windows WebKit2's config.h also includes
WebCore/config.h, which in turn requires to have WebCore/platform in
your include search path. That in turn means that any component that
wants to use WebKit2 headers also needs the WebCore headers, but only
one Windows. It turns out that for the Qt build we don't need it, so
skip it in our configuration and make the include path requirement
the same across all the platforms we build.

* config.h:

Identifier: 114545@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128459 268f45cc-cd09-0410-ab3c-d52691b4dbfc