percentage heights in quirks mode with auto-sized body are computed incorrectly https://bugs.webkit.org/show_bug.cgi?id=96426
Reviewed by Tony Chang.

Source/WebCore:

In quirks mode, for the body element, we need to keep walking up to the html element
if the body's height is auto. Also, we need to subtract the body/html elements
margin/border/padding so that we properly get the body's content height.

This behavior matches IE9, FF and Opera.

None of this applies to standards mode since we don't walk up the
containingBlock chain for auto-sized elements.

Test: fast/css/percentage-height-auto-sized-body-quirks.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::computePercentageLogicalHeight):

LayoutTests:

* fast/css/percentage-height-auto-sized-body-quirks-expected.txt: Added.
* fast/css/percentage-height-auto-sized-body-quirks.html: Added.
Verified that this matches IE9, FF15 and Opera12.

* fast/writing-mode/percentage-height-orthogonal-writing-modes-quirks-expected.txt:
* platform/chromium-linux/fast/table/height-percent-test-expected.png:
* platform/chromium-linux/tables/mozilla/core/cell_heights-expected.png:
* platform/chromium-linux/tables/mozilla_expected_failures/bugs/bug19526-expected.png:
* platform/chromium-linux/tables/mozilla_expected_failures/bugs/bug85016-expected.png:
* platform/chromium-win/tables/mozilla/core/cell_heights-expected.txt:
* platform/chromium-win/tables/mozilla_expected_failures/bugs/bug19526-expected.txt:
* platform/chromium-win/tables/mozilla_expected_failures/bugs/bug85016-expected.txt:
* platform/mac/fast/table/height-percent-test-expected.txt:
Verified that all the changes match FF15.

Identifier: 114603@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128517 268f45cc-cd09-0410-ab3c-d52691b4dbfc