REGRESSION(r128575): fast/forms/time-multiple-fields/time-multiple-fields-preserve-value-after-history-back.html is failed on Chromium-Win https://bugs.webkit.org/show_bug.cgi?id=96754
Reviewed by Kent Tamura.

The root cause of this failure is Chromium-Win port has a second field
event if step=60 due by Windows platform API.

This patch changes test sample to have a second field to get rid of
such difference.

* fast/forms/time-multiple-fields/time-multiple-fields-preserve-value-after-history-back-expected.txt: Updates expectation.
* fast/forms/time-multiple-fields/time-multiple-fields-preserve-value-after-history-back.html: Changed to add "step=1" to
test sample and to add a second field value to expectation.

Identifier: 114677@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@128591 268f45cc-cd09-0410-ab3c-d52691b4dbfc