[Chromium] DRT does not support --dump-all-pixels flag https://bugs.webkit.org/show_bug.cgi?id=95098
Reviewed by Dirk Pranke.

Add support for the --pixel-tests and shorthand -p option in Chromium DumpRenderTree. Use
of this flag causes pixel results to be created for all tests, regardless of
individual test options. If an individual test provides a pixel hash it will be used,
otherwise the hash will be empty. This replaces a previously defined but unused option
--dump-all-pixels, and is useful primarily when debugging DRT instances.

* DumpRenderTree/chromium/DumpRenderTree.cpp:
(runTest): Add a parameter and code to force pixel results for the test.
(main): Add parameter handling for --pixels-test and -p, and remove --dump-all-pixels.


Identifier: 115188@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129139 268f45cc-cd09-0410-ab3c-d52691b4dbfc