-webkit-clip-path is applied on elements that are not descendant of the container https://bugs.webkit.org/show_bug.cgi?id=97217
Reviewed by Dirk Schulze.

Source/WebCore:

The clip-path was set on the GraphicsContext, but was never restored, thus making all the layers
rendered in the same "group" of save/restore state use the same clip-path.

Test: css3/masking/clip-path-restore.html

* css/StyleResolver.cpp:
(WebCore::StyleResolver::collectMatchingRulesForList):
clip-path property should create a stacking-context, otherwise the RenderLayers will not be nested,
meaning that the clip-path of the parent is not going to apply correctly.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintLayerContents):

LayoutTests:

Added a test to check that the clip-path is removed from the GraphicsContext in the second paint call.
clip-path-circle-relative-overflow had incorrect result before, so I've udpated the results on Mac
and added test expectations for the others.

* css3/masking/clip-path-restore-expected.html: Added.
* css3/masking/clip-path-restore.html: Added.
* platform/chromium/TestExpectations:
* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/mac/css3/masking/clip-path-circle-relative-overflow-expected.png:
* platform/qt/TestExpectations:
* platform/win/Skipped:


Identifier: 115261@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129215 268f45cc-cd09-0410-ab3c-d52691b4dbfc